Split flutter samples in different categories
